[@visx/Drag] Allow restriction of drag area
Add object property `restrict` to useDrag (and <Drag>) to set a bounds
for what space the dragged object can occupy.

Min and max values can be set in the x and y plane (or left undefined
for unbounded). This restricts the centre of the object to these bounds.

This works with both the `resetOnStart` and `snapToPointer` options.