Performance Improvements
Calculate and cache sample points along restrictToPath.

This turns out to be fairly tricky as the path is a ref and we need to
use the getTotalLength function as an estimate to when the path has changed.

Remove getParentSvg. Turns out we can just use the DOMMatrix from the
path element itself, saving the need to traverse the DOM.

Fix bug with summing points for offset, rather than finding the difference.